requirements.txt
streamlit
pandas
numpy
matplotlib
openai
reportlab
arabic-reshaper
python-bidi
openpyxl
